v0.0.12
Changed default player for the game to an actor with no collision and no mesh. No movement currently integrated. Added the ability to look around (based on the Z axis) while holding down the right mouse button. Need to fix where the camera is position and include the ability to zoom in and out.