v0.0.14
TowerDefensePlayer:
Changed TowerDefensePlayer from a class of type Pawn to class Character for movement implementation. The physical character is placed under the visible play area with the camera raised up. The player can move and look around the playable area, but cannot see the player.

Edited some properties with the navigation mesh. The navigation mesh *should* update dynamically in game. This is not yet tested.

Enemies:
Changed the enemy collision properties. When the enemies start moving, they will overlap with each other. this allows enemies with different speeds to pass one another instead of get stuck on each other. Added a line of code that places all instances of enemy into a folder labeled "enemies" in the world outliner for organization.