v0.0.22
Added money and lives functionality. Money increases on killing enemies and decreases on building towers.

Round spawning is functional.

Created better UI for the purchase menu for buying a tower.

Removed some deprecated classes no longer being used.

When enemies reach the end of the spline, lives are decreased and the enemy is destroyed.