v0.0.23
Mostly worked on UI. Updated the looks of the user interface and added tool tips to some items that need it. Text dynamically updates in the widget based on the selected tower as well as images. Created the basic layout for adding in upgradeability to towers for damage and range.

Changed how spawning towers works. Now passing a class from blueprints to the C++ function to spawn a tower rather than an integer to get which tower to spawn from a list of classes that inherit from ADefaultTower.