v0.1.5
Updated a lot of the UI graphics.
Fixed issue with crash when particle timer should be expiring.
Added in the ability to open a pause menu.
The pause menu does not really pause the game but gives the player the option to return to the menu or quit.
Began work on the settings menu.